requirements.txt
numpy
fastapi 
uvicorn 
pandas 
scikit-learn 
xgboost 
joblib 
pydantic 
openai 
transformers
python-dotenv
